2016.05.06_lesson21/homework/requirements.txt
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.9
Django==1.9.6
django-filter==0.13.0
djangorestframework==3.3.3
Flask==0.10.1
ipython==4.2.0
ipython-genutils==0.1.0
itsdangerous==0.24
Jinja2==2.8
Markdown==2.6.6
MarkupSafe==0.23
pathlib2==2.1.0
pexpect==4.0.1
pickleshare==0.7.2
ptyprocess==0.5.1
requests==2.10.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.1
Werkzeug==0.11.9
